refactor test file structures
